make loeadbalance example work (at least the times I tested it). It still seems to be required to put something like 14 in alugrid.cfg (4 still give a unpredictable MPI error). Possibly this version also will give an unpredictable MPI error - but unfortunately it is unpredictable...)
Former-commit-id: a0a7dc0a43205ec25479cc93586e752ecf263450

+++ dune/alugrid/common/hsfc.hh
@@ -54,7 +54,7 @@ namespace Dune {
     // return unique hilbert index in interval [0,1] given an element's center
     double hilbertIndex( const Coordinate& point ) const 
     {
-      assert( point.size() == dimension );
+      assert( point.size() == (unsigned int)dimension );
 
       Coordinate center ;
       // scale center into [0,1]^3 box which is needed by Zoltan_HSFC_InvHilbert3d
